+++ .cspell.json
@@ -60,6 +60,7 @@
     "nrwl",
     "paralleldrive",
     "Plutomi",
+    "plutomiid",
     "plutomiinc",
     "praguru",
     "roadmap",
