+++ packages/api/src/controllers.rs
@@ -1,4 +1,5 @@
 pub mod health_check;
+pub mod method_not_allowed;
 pub mod not_found;
 pub mod totp;
