+++ packages/api/src/utils.rs
@@ -1,5 +1,9 @@
 pub mod generate_plutomiid;
 pub mod get_current_time;
 pub mod get_env;
+pub mod headers_to_hashmap;
+pub mod log_req_res;
 pub mod logger;
 pub mod mongodb;
+pub mod parse_request;
+pub mod parse_response;
